refactor: Update ContactForm language based on query parameter
Update the ContactForm component in ContactForm.tsx to dynamically change the language based on the "eng" query parameter. If the "eng" parameter is set to "true", the form will display in English, otherwise it will display in Spanish. This enhancement improves the user experience by allowing users to select their preferred language for the contact form.