refactor: Update ContactForm to include language selection based on query parameter
Improve the ContactForm component in ContactForm.tsx to dynamically change the language based on the "eng" query parameter. If the "eng" parameter is set to "true", the form will display in English; otherwise, it will display in Spanish. This enhancement allows users to select their preferred language for the contact form, improving the overall user experience.